requirements.txt
django==4.2.5
django-environ==0.11.2
environs==9.5.0
loguru==0.7.2
openai==0.28.1
pre-commit==3.4.0
psycopg2-binary==2.9.8   # https://github.com/psycopg/psycopg2
pytest-cov==4.1.0 #https://github.com/pytest-dev/pytest-cov
pytest-django==4.5.2
pytest-emoji==0.2.0
pytest-html==4.0.2  #https://github.com/pytest-dev/pytest-html
pytest-loguru==0.2.0
pytest-md==0.2.0
pytest-mock==3.11.1
pytest-sugar==0.9.7  # https://github.com/Frozenball/pytest-sugar
pytest-xdist==3.3.1
python-dotenv==1.0.0
requests==2.31.0
telethon==1.30.3 #https://github.com/LonamiWebs/Telethon
typing-extensions==4.8.0
watchdog==3.0.0 #https://github.com/gorakhargosh/watchdog
